chore(vite): add missing types and move out utils
