+++ packages/vite/src/utils.ts
@@ -1,0 +1,16 @@
+import { createHash } from 'crypto'
+
+export function hashId (id: string) {
+  return '$id_' + hash(id)
+}
+
+export function hash (input: string, length = 8) {
+  return createHash('sha256')
+    .update(input)
+    .digest('hex')
+    .substr(0, length)
+}
+
+export function uniq<T> (arr: T[]): T[] {
+  return Array.from(new Set(arr))
+}
